+++ examples/options.js
@@ -1,8 +1,8 @@
-const { fn, json } = require('../src');
-const exampleUtils = require('./example-utils');
+const { convertFn } = require('../src');
+const exampleUtils = require('./__example-utils');
 
-const queryUserInfo = fn(exampleUtils.queryUserInfo);
-const queryAccountInfo = fn(exampleUtils.queryAccountInfo, {
+const queryUserInfo = convertFn(exampleUtils.queryUserInfo);
+const queryAccountInfo = convertFn(exampleUtils.queryAccountInfo, {
   debug: true,
   startTime: (name, desc) => {
     console.log('startTime for %s', name);
@@ -12,6 +12,25 @@ const queryAccountInfo = fn(exampleUtils.queryAccountInfo, {
   },
 });
 
-(async () => {
-  console.log(await queryAccountInfo(queryUserInfo()))
-})();
+queryAccountInfo(queryUserInfo()).then(console.log)
+
+/*
+output
+-----------------------------------
+
+startTime for undefined
+airsync: Resolving <unnamed>: {
+  "preferences": 32,
+  "permissions": 7,
+  "user": {
+    "name": "John",
+    "dob": "25/03/1986"
+  }
+}
+endTime for undefined
+{
+  preferences: 32,
+  permissions: 7,
+  user: { name: 'John', dob: '25/03/1986' }
+}
+*/