+++ examples/spread-operator.js
@@ -1,12 +1,8 @@
-// ATTENTION!!
-// jjson and spread are still under development
-// they should not be used in production
+const { resolve, spread, convertFn } = require('../src');
+const exampleUtils = require('./__example-utils');
 
-const { json, spread, fn } = require('../src');
-const exampleUtils = require('./example-utils');
-
-const queryUserInfo = fn(exampleUtils.queryUserInfo);
-const queryAccountInfo = fn(exampleUtils.queryAccountInfo);
+const queryUserInfo = convertFn(exampleUtils.queryUserInfo);
+const queryAccountInfo = convertFn(exampleUtils.queryAccountInfo);
 
 const accountInfo = queryAccountInfo(queryUserInfo());
 
@@ -16,13 +12,24 @@ const other = async () => {
   }
 }
 
-(async () => {
-  const r = await json({
-    accountInfo,
-    [spread()]: queryAccountInfo(queryUserInfo()),
-    [spread()]: other(),
-  });
+resolve({
+  accountInfo,
+  [spread()]: queryAccountInfo(queryUserInfo()),
+  [spread()]: other(),
+}).then(console.log)
 
-  console.log(r);
-
-})();
+/*
+output
+-----
+{
+  accountInfo: {
+    preferences: 32,
+    permissions: 7,
+    user: { name: 'John', dob: '25/03/1986' }
+  },
+  preferences: 32,
+  permissions: 7,
+  user: { name: 'John', dob: '25/03/1986' },
+  name: 'abumq'
+}
+*/