+++ src/get.js
@@ -1,0 +1,6 @@
+const lodashGet = require('lodash.get'); // peer depedency
+const { fn } = require('./fn');
+
+const get = (objOrPromise, path, defaultVal, options = {}) => fn(o => lodashGet(o, path) || defaultVal, options)(objOrPromise)
+
+module.exports.get = get;